ハーベイの公式の組み込み

+++ src/routes/route.py
@@ -1,6 +1,7 @@
 from fastapi import (
     APIRouter
 )
+from decimal import Decimal
 
 from app.cruds.distance import calculate_distance
 
@@ -13,6 +14,6 @@ async def root():
 
 
 @router.get("/distance/")
-async def get_distance(start_longitude : int = 0, start_latitude: int = 0, end_longitude: int = 0, end_latitude: int = 0):
+async def get_distance(start_longitude : Decimal = 0, start_latitude: Decimal = 0, end_longitude: Decimal = 0, end_latitude: Decimal = 0):
     calculate_distance_result = calculate_distance(start_longitude, start_latitude, end_longitude, end_latitude)
     return calculate_distance_result
